Moved test factory and added tilemap system and component.
