+++ core/src/com/itemshop/tilemap/TileMapComponent.java
@@ -1,0 +1,8 @@
+package com.itemshop.tilemap;
+
+/**
+ * An entity's tile map component which defines the tiles it has.
+ */
+public class TileMapComponent {
+
+}
