+++ core/src/com/itemshop/tilemap/TileMapSystem.java
@@ -1,0 +1,10 @@
+package com.itemshop.tilemap;
+
+import com.badlogic.ashley.core.EntitySystem;
+
+/**
+ * Handles processing of each entity's tile map component.
+ */
+public class TileMapSystem extends EntitySystem {
+
+}
